+++ src/common/components/global-route-config/global-route-config.tsx
@@ -1,46 +1,16 @@
-import { useState } from 'react';
-import { useNavigate } from 'react-router';
 import { Outlet } from 'react-router';
 
 import Toast from '@/common/components/toast/toast';
+import { useToast } from '@/common/hooks/use-toast';
 import RecentFeedbackModal from '@/features/profile/components/recent-feedback-modal/recent-feedback-modal';
-import { useFeedbackStore, ProcessState } from '@/store/feedback';
 import { useUserStore } from '@/store/user-auth';
 
 import FeedbackStateObserver from '../feedback-state-observer/feedback-state-observer';
 
 /** 전역적으로 사용되는 로직들을 라우팅 최상단에 배치하는 Config 컴포넌트 */
 export default function GlobalRouteConfig() {
-  const navigate = useNavigate();
-
   const { isAuthenticated, userInfo } = useUserStore();
-
-  const { state, feedbackId, changeState } = useFeedbackStore();
-
-  const [toastOpen, setToastOpen] = useState(true);
-
-  const mapProcessStateToToastType = (state: ProcessState) => {
-    switch (state) {
-      case 'PENDING':
-      case 'IN_PROGRESS':
-        return 'feedbackPending';
-      case 'COMPLETE':
-        return 'feedbackSuccess';
-      case 'ERROR':
-        return 'feedbackError';
-      default:
-        return null;
-    }
-  };
-
-  const navigateTotalEvaluationPage = () => {
-    if (state === 'COMPLETE') {
-      changeState('IDLE');
-      navigate(`/total-evaluation/${feedbackId}`);
-    }
-  };
-
-  const toastType = mapProcessStateToToastType(state);
+  const { toastType, toastOpen, setToastOpen, navigateTotalEvaluationPage } = useToast();
 
   return (
     <>
